Map validation error locations correctly
We now try to find the source for the validation error in the user's query document and use that position as the error location.

If we don't find the location (e.g. it is caused by code generated by graphql-weaver), field that originated the endpoint request (field on root type or link field) is used as error location.