Fix converting formatted errors
Previously, we tried to set the 'locations' property of GraphQLError, but this does not work because it is read-only. These locations are probably incorrect (because they likely use a printed version of the document AST), so we just discard them.

In practice, this fixes a bug where errors were incorrectly reported when they did *not* have locations, because then mapErrorLocations short-cuts and leaves the formatted-error-to-error conversion to convertFormattedErrorsToErrors.